Added water in generation.

+++ SnazzCraft/src/engine/world/light-node.cpp
@@ -1,12 +1,12 @@
 #include "snazzcraft-engine/world/world.hpp"
 
-SnazzCraft::World::LightNode::LightNode(int32_t ILightValue, int32_t IX, int32_t IY, int32_t IZ)
+SnazzCraft::World::LightNode::LightNode(int8_t ILightValue, int32_t IX, int32_t IY, int32_t IZ)
     : X(IX), Y(IY), Z(IZ), LightValue(ILightValue)
 {
 
 }
 
-SnazzCraft::World::LightNode::LightNode(int32_t ILightValue, int32_t IPosition[3])
+SnazzCraft::World::LightNode::LightNode(int8_t ILightValue, int32_t IPosition[3])
     : LightValue(ILightValue)
 {
     this->Position[0] = IPosition[0];
